BSIS-3445 Update validateDonor add validation for donations and deferred donor
This commit updates DonationBackingFormValidator.validateDonor and adds the following checks:
1) for all the donations from the current donor, the donation date is not before the next allowed donation date
2) the donor is not currently deferred

The above checks are already done in DonorConstraintChecker.isDonorEligibleToDonate(); this commit adds these checks to the validator so that explicit error codes are returned (which allows a better handling of these errors at the front-end). Previously, an IllegalArgumentException was thrown by DonationCRUDSErvice which was translated into a 500 error with no clear detail on the exact problem.