BSIS-3574: Fix failing tests in InventoryRepositor
Fixed failing tests in InventoryRepositoryTests which were being caused
by elements being returned in the incorrect order. Because the query is
dependent on the id of the element for ordering, a UUID could cause an order
that isn't related to the createdDate. There is a fix in the UUID generator
for tests, but it seems this wasn't working because this test was failing
often on Travis.

This fix removes the checks on the order of the returned results. There
is a separate issue to fix this problem, and this test can be addressed
there. See: BSIS-3032