BSIS-3544: Improvement to AbstractTypeSafeMatcher
Made a small improvement to AbstractTypeSafeMatcher to make it more explicit
that the "expected" class variable is managed by the parent not the subclass.
It's easy to introduce a bug where the subclass overrides expected and then
when a mismatch occurs in a matcher, we see a NullPointerException because
expected in the super class is not defined.

This bug was fixed in DonationViewModelMatcher.