BSIS-3578: Update TestBatchDonationRange validator
Updated TestBatchDonationRangeBackingFormValidator to include a test that
the Donation specified by the to and from DINs are valid Donation. This
is in line with the acceptance criteria for error messages.

Removed the 'noDonations' error as it would no longer be relevant as at
least one Donation is guaranteed to exist in the range.