BSIS-3576: Remove TestBatch from DonationBatch
Removed TestBatch association from DonationBatch because Donations are now
linked directly to TestBatches.

Also removed one query that was no longer being used and wouldn't run
once the association had been removed - the query to find unassigned
donation batches.

Removed the constraint for re-opening a DonationBatch that checked if
there was a TestBatch associated - this constraint is no longer applicable.