BSIS-3610: Partial revert of bd47dbc
Partial revert of bd47dbc where lastUpdated was removed from the view model.
This commit returns the ignored setter for lastUpdated in TestBatchBackingForm
because updating the TestBatch from the frontend was failing due to an unknown
property 'lastUpdated'. This commit also fixes the annoying issue where the
JavaBean property is named differently from the setter. This meant that originally
'lastUpdatedDate' was checked in the frontend and found to not exist.