BSIS-3625: Fix issue with TestBatchCRUDService
Fixed an issue in the TestBatchCRUDService where donations that were added
to the test batch weren't associated correctly which resulted in the donations
not being added to the test batch.

Added a specific "createTestBatch" method to the TestBatchCRUDService which
is more in line with what is done in other CRUD services - so the business
logic code was no longer in the ControllerService and could be tested.