BSIS-3636: Remove isTestingSupervisor boolean
Removed isTestingSupervisor boolean which was being used when determining
constraint permissions. Constraint permissions should just indicate what
actions are possible on the entity in question, not whether the logged in
user is allowed to perform the action. The roles/permissions should be
checked separately (in the frontend and on the rest endpoint).

So, for example, if a user has a permission to Void Test Batch, but the
Test Batch cannot be voided, then they should see a disabled Void link.
But if the user does not have the permission to Void Test Batch, then they
should not see the Void link (it is irrelevant whether the Test Batch can
be voided or not).