BSIS-3726 Ignore discardReason fields in PreProcessingBackingForm
The fields discardReason and discardReasonComment were added to
the ComponentManagementViewModel which was sent down to the
front-end. This same data was used in the preprocess of the component
when recording the weight. However, the ComponentPreProcessingBackingForm
does not expect the new fields.

To handle this the ComponentPreProcessingBackingForm was updated to
ignore all unkown Json fields.