BSIS-3752 Add edit pack type rule to Donation Contraint Checker
Previously the pack type of a donation with a test batch that has been
released could be changed from one that produces test samples to
one that does not.

If a sample (DIN) has been released the user should NOT be able to
edit the pack type IF the pack type is changed from one that produces
 a test sample to one that does not produce a test sample.

This commit adds the rule to the backend.