BSIS-3752 Add per Donation isReleased check to canEditNewPackType
Previously the canEditToNewPackType would return false if the
test Batch was not open, and the pack type change is from one
that produces test samples to one that doesn't or vice versa.

This commit in addition to checking whether the test batch is not open
also caters for when the test batch is open but the donation is marked
as released. This is because Donations may or may not be released
if the test batch is not.

Also refactored the code a bit to be more readable.