BSIS-3775 Fix component preprocess request mapping
Previously there was a trailing space at the end of the
RequestMapping for the components/preprocess endpoint. This
was not a problem before until updating the relevant spring
dependency. After updated the spring dependency the trailing
space caused the API call, without the trailing space, to not
match the request mapping.

The trailing space is removed in this commit